backend/requirements.txt
Flask
Flask-Cors
gunicorn
requests
mysql-connector-python
react-router-dom